Enhance product management: add code field to products and update form with code input; make SKU and barcode optional

+++ database/migrations/2025_08_10_054249_create_sales_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->foreignIdFor(Customer::class)->nullable()->constrained()->nullOnDelete();
             $table->integer('subtotal')->nullable()->comment('Subtotal before tax and discount. Calculated from product_sale');
             $table->integer('tax')->nullable()->comment('Total tax amount. Calculated from product_sale');
-            $table->float('discount')->nullable()->comment('Discount on subtotal');
+            $table->decimal('discount')->nullable()->comment('Discount on subtotal');
             $table->integer('total')->nullable();
             $table->string('status')->default(SaleStatus::Completed);
             $table->string('payment_status')->default(SalePaymentStatus::default());
@@ -34,10 +34,10 @@ return new class extends Migration
         Schema::create('product_sale', function (Blueprint $table) {
             $table->foreignIdFor(Sale::class)->constrained()->cascadeOnDelete();
             $table->foreignIdFor(Product::class)->nullable()->constrained()->nullOnDelete();
-            $table->integer('quantity')->default(1);
+            $table->decimal('quantity')->default(1);
             $table->integer('unit_price');
             $table->integer('tax')->nullable();
-            $table->float('discount')->nullable()->comment('Discount in percent');
+            $table->decimal('discount')->nullable()->comment('Discount in percent');
             $table->integer('supplier_price')->nullable();
         });
     }
